Add files via upload
This is the final version of my Square-Recognition neural network project. I have built my own version of a neural network from scratch and trained it to recognize squares. The entire notebook can be executed in under 5 minutes. My neural network can be trained in just a few minutes. A Jupyter Notebook file is attached here which contains the code that builds, trains, and tests the neural network.